% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_conn}
\alias{grab_conn}
\title{Return a specific connection from the prelude}
\usage{
grab_conn(conn_name)
}
\arguments{
\item{conn_name}{A character string of the connection name.}
}
\value{
A SQL Server connection object.
}
\description{
Return a specific connection from the prelude
}
\seealso{
Other grabs: 
\code{\link{grab_prelude}()},
\code{\link{grab_prelude_auth}()},
\code{\link{grab_prelude_errors}()},
\code{\link{grab_prod_conn}()}
}
\concept{grabs}
