% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_prelude}
\alias{grab_prelude}
\title{Return the prelude object from the "the" environment}
\usage{
grab_prelude()
}
\value{
The prelude object invisibly - a named list from the
\code{the} environment with connections and authentication.
}
\description{
Return the prelude object from the "the" environment
}
\examples{
global_prelude_object <- grab_prelude()
}
\seealso{
Other grabs: 
\code{\link{grab_conn}()},
\code{\link{grab_prelude_auth}()},
\code{\link{grab_prelude_errors}()},
\code{\link{grab_prod_conn}()}
}
\concept{grabs}
