% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_prelude_auth}
\alias{grab_prelude_auth}
\title{Return authentication from the prelude}
\usage{
grab_prelude_auth()
}
\value{
A named list of authentication objects.
}
\description{
Return authentication from the prelude
}
\examples{
grab_prelude_auth()
}
\seealso{
Other grabs: 
\code{\link{grab_conn}()},
\code{\link{grab_prelude}()},
\code{\link{grab_prelude_errors}()},
\code{\link{grab_prod_conn}()}
}
\concept{grabs}
