% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_prelude_errors}
\alias{grab_prelude_errors}
\title{Return captured errors from the prelude}
\usage{
grab_prelude_errors()
}
\value{
A named list of prelude-related errors.
}
\description{
Return captured errors from the prelude
}
\examples{
grab_prelude_errors()
}
\seealso{
Other grabs: 
\code{\link{grab_conn}()},
\code{\link{grab_prelude}()},
\code{\link{grab_prelude_auth}()},
\code{\link{grab_prod_conn}()}
}
\concept{grabs}
