% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_prod_conn}
\alias{grab_prod_conn}
\alias{grab_stg_conn}
\alias{grab_dev_conn}
\title{Return connections from the prelude}
\usage{
grab_prod_conn()

grab_stg_conn()

grab_dev_conn()
}
\value{
A SQL Server connection object.
}
\description{
\code{grab_prod_conn()}returns the production connection.
\code{grab_stg_conn()} returns the staging connection.
\code{grab_dev_conn()} returns the development connection.
}
\examples{
prod <- grab_prod_conn()
stg <- grab_stg_conn()
dev <- grab_dev_conn()
}
\seealso{
Other grabs: 
\code{\link{grab_conn}()},
\code{\link{grab_prelude}()},
\code{\link{grab_prelude_auth}()},
\code{\link{grab_prelude_errors}()}

Other grabs: 
\code{\link{grab_conn}()},
\code{\link{grab_prelude}()},
\code{\link{grab_prelude_auth}()},
\code{\link{grab_prelude_errors}()}

Other grabs: 
\code{\link{grab_conn}()},
\code{\link{grab_prelude}()},
\code{\link{grab_prelude_auth}()},
\code{\link{grab_prelude_errors}()}
}
\concept{grabs}
