% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{make_timestamp}
\alias{make_timestamp}
\title{Create a standard timestamp}
\usage{
make_timestamp()
}
\value{
A character string of a standardized timestamp.
}
\description{
Create a standard timestamp
}
\examples{
make_timestamp()
}
