% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelude.R
\name{prelude}
\alias{prelude}
\title{Setup Standard Environment}
\usage{
prelude()
}
\value{
The prelude object invisibly - a named list from the
\code{the} environment with connections and authentication.
}
\description{
\code{prelude()} sets up the standard connections required, authentication
and other common objects needed for analytics and development and
stores it in the "the" environment.
}
\examples{
prelude()
}
