% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{setup_auth}
\alias{setup_auth}
\title{Grab Relevant Authentication Artifacts from System Keyring}
\usage{
setup_auth(service_name)
}
\arguments{
\item{service_name}{Name of Service in System Keyring.}
}
\value{
An invisible named list of credentials.
}
\description{
\code{setup_auth()} will return a named list including,
most commonly, the email and password related to
the service name provided.
}
\examples{
\dontrun{
auth <- setup_auth("service1")
foo <- function(url, email, password) {
 args <-
   list(
     url,
     email,
     password
 )
 invisible(TRUE)
}

foo(
  url = "https://someurl.com/resource/12315.json",
  email = auth$service1$email,
  password = auth$service1$password
)
}
}
