% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{setup_connections}
\alias{setup_connections}
\title{Setup Sql Server Connections}
\usage{
setup_connections()
}
\value{
An invisible named list of connections.
}
\description{
\code{setup_connections()} checks the system's DSNs for
the expected sql servers and if found
will setup the connections for the user and return
in a named list.  If the expected DSNs are not found
an error will be thrown.
}
\examples{
\dontrun{
connections <- setup_connections()
table <- odp::read_table(
  conn = connections$prod,
  database_name = "fake_db",
  schema_name = "fake_schema",
  table_name = "fake_table",
  lazy = TRUE
)
}
}
